+++ packages/database/src/index.ts
@@ -11,6 +11,7 @@ export * from "./generated/database.types";
 
 export { createStudiosRepository } from "./repositories/studios-repository";
 export { createUsersRepository } from "./repositories/users-repository";
+export { createInvitesRepository } from "./repositories/invites-repository";
 export { createProjectsRepository } from "./repositories/projects-repository";
 export { createGamesRepository } from "./repositories/games-repository";
 export { createKnowledgeDocumentsRepository } from "./repositories/knowledge-documents-repository";